+++ src/contentChunk.js
@@ -44,7 +44,7 @@ exports.extractStringPayloadIntoChunks = (payload) => {
     if (!isValidGridFsPayload(payload)) {
       return reject(new Error('payload not in the correct format, expecting a string, Buffer, ArrayBuffer, Array, or Array-like Object'))
     }
-    
+
     const bucket = new mongodb.GridFSBucket(connectionDefault.client.db())
     const stream = bucket.openUploadStream()
 
@@ -60,4 +60,30 @@ exports.extractStringPayloadIntoChunks = (payload) => {
     })
     stream.end(payload)
   })
+}
+
+export const retrieveTransactionBody = (fileId, callback) => {
+  const db = connectionDefault.client.db()
+  if (!db) {
+      const err = new Error(`Transaction body retrieval failed. Database handle: ${db} is invalid`)
+      return callback(err, null)
+  }
+
+  if (!fileId) {
+    const err = new Error(`Transaction body retrieval failed: Transaction id: ${fileId}`)
+    return callback(err, null)
+  }
+
+  const bucket = new mongodb.GridFSBucket(db)
+
+  let body = ''
+  bucket.openDownloadStream(fileId)
+    .on('error', err => {
+      const error = new Error(`Transaction body retrieval failed: Error in reading stream: ${err.message}`)
+      return callback(error, null)
+    })
+    .on('data', chunk => body += chunk)
+    .on('end', () => {
+      return callback(null, body)
+    })
 }
